Various small changes.
About to update Unification.combine and EffectOn.apply.  At present, I think
it gets secondary induced transitions wrong.  It's necessary to allow
parameters of cv to be remapped to any parameter of the component of post that
gains a reference to cv (at present it's all parameters of all components of
*pre*).  Also, the current implementation is inefficient, with
Unification.combine producing too many results.  It's better to separate out
the secondary induced transitions.